backend/requirements.txt
flask
flask_cors
pymysql
wheel
bcrypt
python-dotenv
PyJWT
flask_sqlalchemy
mysqlclient
Flask-Migrate
flask_jwt_extended